Move split_batches to misc_utils.
